Organize admin components; add header logo
Move admin components into brand/, event/, and member/ subfolders and update imports across pages. Replace placeholder LOGO with an imported logo image in AdminLayout. Remove a redundant comment in events table and adjust a few import paths/extensions to match relocated files.
